git-svn-id: svn://svn.code.sf.net/p/daetools/code@309 e944343b-51ee-4bc0-bd0d-d7880edd40ba

+++ trunk/IDAS_DAESolver/ida_solver.cpp
@@ -27,7 +27,7 @@ int roots(realtype	time,
 		  realtype*	gout, 
 		  void*		pUserData);
 
-int jacobian(int	    Neq, 
+int jacobian(long int    Neq, 
 			 realtype	time, 
 			 realtype	dInverseTimeStep, 
 			 N_Vector	vectorVariables, 
@@ -1017,7 +1017,7 @@ int roots(realtype	time,
 	return 0;
 }
 
-int jacobian(int	    Neq, 
+int jacobian(long int    Neq, 
 			 realtype	time, 
 			 realtype	dInverseTimeStep, 
 			 N_Vector	vectorVariables,
